requirements.txt
-e .
pyaml
pandas
scikit-learn
evidently
